--- orchestration/dagster_orchestration/hca_orchestration/resources/data_repo.py
@@ -1,43 +1,0 @@
-from dataclasses import dataclass
-
-from dagster import resource, StringSource, Field
-from dagster.core.execution.context.init import InitResourceContext
-
-from data_repo_client import ApiClient, Configuration, RepositoryApi
-
-from hca_orchestration.contrib.google import default_google_access_token
-
-
-@resource({
-    "api_url": Field(StringSource)
-})
-def jade_data_repo_client(init_context: InitResourceContext) -> RepositoryApi:
-    # create API client
-    config = Configuration(host=init_context.resource_config["api_url"])
-    config.access_token = default_google_access_token()
-    client = ApiClient(configuration=config)
-    client.client_side_validation = False
-
-    # submit file ingest (for now just enumerate datasets or something to prove interaction works)
-    return RepositoryApi(api_client=client)
-
-
-class NoopDataRepoClient:
-    @dataclass
-    class NoopResult:
-        total: int
-
-    @dataclass
-    class FakeJobResponse:
-        completed: bool
-
-    def enumerate_datasets(self) -> NoopResult:
-        return NoopDataRepoClient.NoopResult(5)
-
-    def retrieve_job(self, job_id: str) -> FakeJobResponse:
-        return NoopDataRepoClient.FakeJobResponse(True)
-
-
-@resource
-def noop_data_repo_client(init_context: InitResourceContext) -> NoopDataRepoClient:
-    return NoopDataRepoClient()
