+++ orchestration/dagster_orchestration/hca_orchestration/solids/load_hca/stage_data.py
@@ -1,6 +1,6 @@
 import re
 
-from dagster import solid, InputDefinition, Nothing, String, Int
+from dagster import solid, InputDefinition, Nothing, String
 from dagster.core.execution.context.compute import AbstractComputeExecutionContext
 from google.cloud.bigquery import Dataset
