+++ orchestration/dagster_orchestration/hca_orchestration/tests/solids/test_data_repo.py
@@ -2,9 +2,9 @@ import unittest
 from unittest.mock import Mock, patch
 
 from dagster import execute_solid, Failure, ModeDefinition
+from dagster_utils.resources.data_repo import noop_data_repo_client
 
 from hca_manage.manage import JobId
-from hca_orchestration.resources.data_repo import noop_data_repo_client
 from hca_orchestration.solids.data_repo import base_wait_for_job_completion
 
 
@@ -46,7 +46,7 @@ class WaitForJobCompletionTestCase(unittest.TestCase):
             mock_job_status(True)
         ]
 
-        with patch('hca_orchestration.resources.data_repo.NoopDataRepoClient.retrieve_job',
+        with patch('dagster_utils.resources.data_repo.NoopDataRepoClient.retrieve_job',
                    side_effect=job_status_sequence) as mocked_retrieve_job:
             result = execute_solid(
                 base_wait_for_job_completion,
