+++ orchestration/dagster_orchestration/hca_orchestration/tests/solids/test_validate_egress.py
@@ -2,14 +2,13 @@ import unittest
 from unittest.mock import Mock, patch
 
 from dagster import execute_solid
+from dagster_utils.testing.matchers import StringContaining
 
 from hca_manage.manage import ProblemCount
 
 from hca_orchestration.pipelines.validate_egress import test_mode
 from hca_orchestration.solids.validate_egress import post_import_validate,\
     notify_slack_of_egress_validation_results
-
-from hca_orchestration.tests.support.matchers import StringContaining
 
 
 class PostImportValidateTestCase(unittest.TestCase):
