+++ orchestration/dagster_orchestration/pyproject.toml
@@ -5,7 +5,7 @@ description = ""
 authors = ["Your Name <you@example.com>"]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "3.9"
 argo-workflows = "^5.0.0"
 cached-property = "^1.5.2"
 dagster = "0.11.6"
@@ -18,6 +18,7 @@ google-cloud-storage = "^1.3.5"
 python-dateutil = "^2.8.1"
 typing-extensions = "^3.7.4"
 pyyaml = "^5.3"
+broad-dagster-utils = "^0.1.2"
 
 [tool.poetry.dev-dependencies]
 autopep8 = "^1.5.5"
